艺术家接口

+++ app/src/main/java/com/bohui/art/bean/detail/ArtManLevelBean.java
@@ -10,19 +10,19 @@ import java.io.Serializable;
 
 
 public class ArtManLevelBean implements Serializable {
-    private long id;
+    private int id;
     private String name;
 
-    public ArtManLevelBean(long id, String name) {
+    public ArtManLevelBean(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
